Open file interface and shortcut keys
Integrates the process of opening a file into the File menu like any
normal program would and sets up standard shortkey keys for open, save,
and save as.

Also displays the loaded filename as part of the application title so
that you can tell which file is which in the task bar even if you have
several copies of WeaponEd open.